+++ pkger/http_server_test.go
@@ -857,6 +857,10 @@ func (f *fakeSVC) DeleteStack(ctx context.Context, identifiers struct{ OrgID, Us
 	panic("not implemented yet")
 }
 
+func (f *fakeSVC) ExportStack(ctx context.Context, orgID, stackID influxdb.ID) (*pkger.Pkg, error) {
+	panic("not implemented")
+}
+
 func (f *fakeSVC) ListStacks(ctx context.Context, orgID influxdb.ID, filter pkger.ListFilter) ([]pkger.Stack, error) {
 	if f.listStacksFn == nil {
 		panic("not implemented")